+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+[Unreleased]
+
+## [4.0.0](https://github.com/nfriedly/node-bestzip/releases/tag/v4.0.0) - 2026-10-06
+
+### Security
+
+- Default to not following symlinks to prevent arbitrary file read ([GHSA-rq7v-fpj4-v4vc](https://github.com/nfriedly/node-bestzip/security/advisories/GHSA-rq7v-fpj4-v4vc))
+
+### Breaking
+
+- Default behavior of symlinks changed from including the target to including the link. Set `--follow-sym-links` or `followSymLinks: true` to restore previous behavior.
+
 
 ## [3.1.0](https://github.com/nfriedly/node-bestzip/releases/tag/v3.1.0) - 2026-09-06